requirements.txt
requests>=2.31
Pillow>=10.0
jsonschema>=4.21
